autonomous-ops/src/main/python/ml/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.3
python-multipart==0.0.6
grpcio==1.59.0
grpcio-tools==1.59.0
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.5.1
tensorflow==2.16.1
pytorch==2.2.0
pandas==2.1.4
matplotlib==3.8.4
seaborn==0.13.2